separate file for commandParser class
